backend/requirements.txt
# Flask 核心框架
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# 数据库
SQLAlchemy==2.0.23

# HTTP 请求
requests==2.31.0
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
Pillow==10.1.0

# 阿里云百炼 SDK
dashscope==1.14.1
